server/pyproject.toml
[project]
name = "ai-chatter"
version = "0.1.0"
description = "AiChatter - M5StickS3 常時会話AIアシスタント サーバー"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "mlx-audio[stt,tts]>=0.3.1",
    "openai>=1.60.0",
    "numpy>=1.26.0",
    "pydantic-settings>=2.2.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.13.0",
    "pyyaml>=6.0",
    "pyserial>=3.5",
    "esptool>=4.0",
    "httpx>=0.27.0",
    "sudachipy>=0.6.8",
    "sudachidict_core>=20240716",
    "rank_bm25>=0.2.2",
    "resvg-py>=0.2.6",
    "transformers>=4.30.0",
    "torch>=2.0.0",
]

[project.scripts]
ai-chatter = "ai_chatter.cli:main"

[build-system]
requires = ["uv_build>=0.7.2,<0.8"]
build-backend = "uv_build"
